+++ src/factory/collections/factory_vec.rs
@@ -38,6 +38,16 @@ where
         }
     }
 
+    /// Returns the length as amount of elements stored in this type.
+    pub fn len(&self) -> usize {
+        self.data.len()
+    }
+
+    /// Returns [`true`] if the length of this type is `0`.
+    pub fn is_empty(&self) -> bool {
+        self.data.is_empty()
+    }
+
     /// Insert an element at the end of a [`FactoryVec`].
     pub fn push(&mut self, data: Data) {
         let index = self.data.len();
@@ -62,19 +72,19 @@ where
     }
 
     /// Get a reference to data stored at `index`.
-    pub fn get(&self, index: usize) -> &Data {
-        &self.data[index]
+    pub fn get(&self, index: usize) -> Option<&Data> {
+        self.data.get(index)
     }
 
     /// Get a mutable reference to data stored at `index`.
     ///
     /// Assumes that the data will be modified and the corresponding widget
     /// needs to be updated.
-    pub fn get_mut(&mut self, index: usize) -> &mut Data {
+    pub fn get_mut(&mut self, index: usize) -> Option<&mut Data> {
         let mut changes = self.changes.borrow_mut();
         changes.entry(index).or_insert(ChangeType::Update);
 
-        &mut self.data[index]
+        self.data.get_mut(index)
     }
 }