+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
++ Book: Factory chapters
+
+### Changed
+
++ get and get_mut of FactoryVec and FactoryVecDeque now return an Option to prevent panics
+
 ### Fixed
 
 + Fix unsound removal of elements in FactoryVecDeque